separate logic level implementation of taskqueue from platform
feedback_services contains function to add task in taskqueue, which
calls the taskqueue service to add particular task in queue.
This separation provides interface between function which communicates with GAE
and function which is used by other code.
